src/containers/ucdp-container/requirements.txt
azure-functions
azure-storage-blob
requests
pandas
pyarrow